improved q1qvm support - now runs on 64bit servers with 32bit qvm
git-svn-id: https://svn.code.sf.net/p/fteqw/code/trunk@2635 fc73d0e0-1445-4013-8a0c-d673dee63da5

+++ engine/common/vm.h
@@ -11,12 +11,12 @@
 
 
 typedef int (EXPORT_FN *sys_calldll_t) (int arg, ...);
-typedef long (*sys_callqvm_t) (void *offset, unsigned int mask, int fn, const long *arg);
+typedef int (*sys_callqvm_t) (void *offset, unsigned int mask, int fn, const int *arg);
 
 typedef struct vm_s vm_t;
 
 // for syscall users
-#define VM_LONG(x)		(*(long*)&(x))
+#define VM_LONG(x)		(*(int*)&(x))
 #define VM_FLOAT(x)		(*(float*)&(x))
 #define VM_POINTER(x)	((x)?(void*)((char *)offset+((x)%mask)):NULL)
 #define VM_OOB(p,l)		(p + l >= mask || VM_POINTER(p) < offset)
